+++ app/app.rb
@@ -10,46 +10,6 @@ class App
     @books_list = []
     @people_list = []
     @rentals_list = []
-
-    loop do
-      show_menu
-      user_choice = gets.chomp
-      handle_option(user_choice)
-    end
-  end
-
-  def show_menu
-    puts
-    puts 'Please choose an option by entering a number:'
-    puts '1 - List all books'
-    puts '2 - List all people'
-    puts '3 - Create a person'
-    puts '4 - Create a book'
-    puts '5 - Create a rental'
-    puts '6 - List all rentals for a given person id'
-    puts '7 - Exit'
-    puts
-  end
-
-  def handle_option(user_choice)
-    case user_choice
-    when '1'
-      list_all_books
-    when '2'
-      list_all_people
-    when '3'
-      create_person
-    when '4'
-      create_book
-    when '5'
-      create_rental
-    when '6'
-      list_all_rentals
-    when '7'
-      exit_app
-    else
-      puts 'Invalid option, please try again!'
-    end
   end
 
   def list_all_books
@@ -151,10 +111,5 @@ class App
       puts "Date: #{rental.date}, Book \"#{rental.book.title}\" by #{rental.book.author}" if rental.person.id == id
     end
     @parent.show_menu
-  end
-
-  def exit_app
-    puts 'Thank you! Great Day!'
-    exit(0)
   end
 end
