+++ src/person.rb
@@ -3,7 +3,7 @@ require_relative 'capitalizedecorator'
 require_relative 'trimmerdecorator'
 
 class Person < Nameable
-  attr_reader :id, :rentals
+  attr_reader :id
   attr_accessor :name, :age
 
   def initialize(age, name = 'Unknown', parent_permission: true)
@@ -12,7 +12,6 @@ class Person < Nameable
     @name = name
     @age = age
     @parent_permission = parent_permission
-    @rentals = []
   end
 
   def correct_name
@@ -27,9 +26,5 @@ class Person < Nameable
 
   def generate_id
     rand(1..1000)
-  end
-
-  def add_rental(rental)
-    @rentals << rental
   end
 end
