+++ src/student.rb
@@ -4,7 +4,7 @@ class Student < Person
   attr_reader :classroom
 
   def initialize(age, classroom, name = 'Unknown', parent_permission: true)
-    super(age, name)
+    super(age, name, parent_permission: parent_permission)
     @classroom = classroom
   end
